more simplification and pruning
