--- bokeh/server/views/deps.py
@@ -1,13 +1,0 @@
-#-----------------------------------------------------------------------------
-# Copyright (c) 2012 - 2015, Continuum Analytics, Inc. All rights reserved.
-#
-# Powered by the Bokeh Development Team.
-#
-# The full license is in the file LICENSE.txt, distributed with this software.
-#-----------------------------------------------------------------------------
-from __future__ import absolute_import
-
-from . import backbone, main, plugins, statics
-
-# this just shuts up pyflakes
-backbone, main, plugins, statics
